Rename spanstore To tracestore
