+++ storage_v2/tracestore/factory.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package spanstore
+package tracestore
 
 import (
 	"github.com/jaegertracing/jaeger/storage_v2"
