+++ storage_v2/tracestore/reader.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package spanstore
+package tracestore
 
 import (
 	"context"
